Added swedate functions: * swe_utc_time_zone * swe_utc_to_jd
Notice: swe_utc_to_jd uses the swe_deltat_ex function, but we made file reading working, so it's not a problem at all to read this file. But the code base of that function is a bit complex, so in next step I will implement that function as priority.